Improve structure and documentation of Multi-index Hits connector
